Move format_benchmark_result to asv.results
